this was kinda dumb
I mean this doesn't even consider any other
subdomain other than 'www'. Like that's the only
subdomain that exists.

I don't mind it being the only subdomain for this challenge
but it is kinda silly that they don't really mention anything
about the lack of reality that this kata is.